Expand the README
In particular, add a note that setup instructions are in the app itself,
since they're kind of deemphasized in the UI. This confused me for a
moment when I didn't notice them, went looking for instructions in the
README instead, and found nothing.

Installation instructions shamelessly cribbed from nextcloud/photos.

Signed-off-by: AJ Jordan <alex@strugee.net>